Change to_i to to_f type

+++ task1.rb
@@ -1,7 +1,7 @@
 print "Привет, как тебя звать? "
 name = gets.chomp.capitalize!
 print "Какой у вас рост?(см) "
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 weight = height - 110
 if weight <= 0
   puts "#{name}, твой вес оптимальный"
