+++ task3.rb
@@ -1,9 +1,9 @@
 print "Введите длину 1 стороны треугольника "
-length1 = gets.chomp.to_i
+length1 = gets.chomp.to_f
 print "теперь длину 2 стороны треугольника "
-length2 = gets.chomp.to_i
+length2 = gets.chomp.to_f
 print "и длину 3 стороны треугольника "
-length3 = gets.chomp.to_i
+length3 = gets.chomp.to_f
 max = [ length1, length2, length3].max(3)
 if max[0] ** 2 == max[1] ** 2 + max[2] ** 2
   rectangular_triangle = true
